process to add events
