+++ package.json
@@ -6,9 +6,11 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.2.2",
     "date-fns": "^2.29.3",
     "react": "^18.2.0",
     "react-big-calendar": "^1.5.2",
+    "react-bootstrap": "^2.6.0",
     "react-datepicker": "^4.8.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
